chore: fix LanguageContext fast refresh warning
by moving React context to a separate file
